requirements.txt
aiogram==3.2.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
aiosqlite==0.19.0
httpx==0.25.1
pydantic==2.5.0
